pyproject.toml
[project]
authors = [
    {name = "Ivan Foksha", email = "v.foksh6@gmail.com"}
]
name = "hexlet-code"
version = "0.1.0"
description = "Analyze your or own website on the Internet"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.4",
    "validators>=0.35.0",
    "requests>=2.32.4",
]

[build-system] 
requires = ["hatchling"] 
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel] 
packages = ["page_analyzer"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["I001"]
